+++ common/CMakeLists.txt
@@ -88,7 +88,6 @@ target_sources(common PRIVATE
 	RwMutex.h
 	SafeArray.h
 	ScopedGuard.h
-	ScopedPtrMT.h
 	SettingsInterface.h
 	SettingsWrapper.h
 	StringHelpers.h
