+++ common/Threading.h
@@ -15,6 +15,7 @@
 
 #pragma once
 
+#include <wx/datetime.h>
 #ifdef _WIN32
 // thanks I hate it.
 #include <wx/filefn.h>
@@ -30,43 +31,6 @@
 #include "common/TraceLog.h"
 #include "common/General.h"
 #include <atomic>
-
-#undef Yield // release the burden of windows.h global namespace spam.
-
-#define AffinityAssert_AllowFrom_MainUI() \
-	pxAssertMsg(wxThread::IsMain(), "Thread affinity violation: Call allowed from main thread only.")
-
-// --------------------------------------------------------------------------------------
-//  pxThreadLog / ConsoleLogSource_Threading
-// --------------------------------------------------------------------------------------
-
-class ConsoleLogSource_Threading : ConsoleLogSource
-{
-	typedef ConsoleLogSource _parent;
-
-public:
-	using _parent::IsActive;
-
-	ConsoleLogSource_Threading();
-
-	bool Write(const wxString& thrname, const wxChar* msg)
-	{
-		return _parent::Write(wxsFormat(L"(thread:%s) ", WX_STR(thrname)) + msg);
-	}
-	bool Warn(const wxString& thrname, const wxChar* msg)
-	{
-		return _parent::Warn(wxsFormat(L"(thread:%s) ", WX_STR(thrname)) + msg);
-	}
-	bool Error(const wxString& thrname, const wxChar* msg)
-	{
-		return _parent::Error(wxsFormat(L"(thread:%s) ", WX_STR(thrname)) + msg);
-	}
-};
-
-extern ConsoleLogSource_Threading pxConLog_Thread;
-
-#define pxThreadLog pxConLog_Thread.IsActive() && pxConLog_Thread
-
 
 // --------------------------------------------------------------------------------------
 //  PCSX2_THREAD_LOCAL - Defines platform/operating system support for Thread Local Storage
@@ -86,8 +50,6 @@ extern ConsoleLogSource_Threading pxConLog_Thread;
 #define DeclareTls(x) x
 #endif
 
-class wxTimeSpan;
-
 namespace Threading
 {
 	class ThreadHandle;
@@ -95,74 +57,16 @@ namespace Threading
 	class RwMutex;
 
 	extern void pxTestCancel();
-	extern pxThread* pxGetCurrentThread();
-	extern wxString pxGetCurrentThreadName();
+	extern void YieldToMain();
+
 	extern u64 GetThreadCpuTime();
 	extern u64 GetThreadTicksPerSecond();
 
 	/// Set the name of the current thread
 	extern void SetNameOfCurrentThread(const char* name);
 
-	// Yields the current thread and provides cancellation points if the thread is managed by
-	// pxThread.  Unmanaged threads use standard Sleep.
-	extern void pxYield(int ms);
+	extern const wxTimeSpan def_yieldgui_interval;
 } // namespace Threading
-
-namespace Exception
-{
-	class BaseThreadError : public RuntimeError
-	{
-		DEFINE_EXCEPTION_COPYTORS(BaseThreadError, RuntimeError)
-		DEFINE_EXCEPTION_MESSAGES(BaseThreadError)
-
-	public:
-		Threading::pxThread* m_thread;
-
-	protected:
-		BaseThreadError()
-		{
-			m_thread = NULL;
-		}
-
-	public:
-		explicit BaseThreadError(Threading::pxThread* _thread)
-		{
-			m_thread = _thread;
-			m_message_diag = L"An unspecified thread-related error occurred (thread=%s)";
-		}
-
-		explicit BaseThreadError(Threading::pxThread& _thread)
-		{
-			m_thread = &_thread;
-			m_message_diag = L"An unspecified thread-related error occurred (thread=%s)";
-		}
-
-		virtual wxString FormatDiagnosticMessage() const;
-		virtual wxString FormatDisplayMessage() const;
-
-		Threading::pxThread& Thread();
-		const Threading::pxThread& Thread() const;
-	};
-
-	class ThreadCreationError : public BaseThreadError
-	{
-		DEFINE_EXCEPTION_COPYTORS(ThreadCreationError, BaseThreadError)
-
-	public:
-		explicit ThreadCreationError(Threading::pxThread* _thread)
-		{
-			m_thread = _thread;
-			SetBothMsgs(L"Thread creation failure.  An unspecified error occurred while trying to create the %s thread.");
-		}
-
-		explicit ThreadCreationError(Threading::pxThread& _thread)
-		{
-			m_thread = &_thread;
-			SetBothMsgs(L"Thread creation failure.  An unspecified error occurred while trying to create the %s thread.");
-		}
-	};
-} // namespace Exception
-
 
 namespace Threading
 {
@@ -185,22 +89,6 @@ namespace Threading
 
 	// sleeps the current thread for the given number of milliseconds.
 	extern void Sleep(int ms);
-
-// pthread Cond is an evil api that is not suited for Pcsx2 needs.
-// Let's not use it. Use mutexes and semaphores instead to create waits. (Air)
-#if 0
-	struct WaitEvent
-	{
-		pthread_cond_t cond;
-		pthread_mutex_t mutex;
-
-		WaitEvent();
-		~WaitEvent();
-
-		void Set();
-		void Wait();
-	};
-#endif
 
 	// --------------------------------------------------------------------------------------
 	//  ThreadHandle
@@ -374,48 +262,6 @@ namespace Threading
 		void Reset();
 	};
 
-	class Semaphore
-	{
-	protected:
-#ifdef __APPLE__
-		semaphore_t m_sema;
-		int m_counter;
-#else
-		sem_t m_sema;
-#endif
-
-	public:
-		Semaphore();
-		virtual ~Semaphore();
-
-		void Reset();
-		void Post();
-		void Post(int multiple);
-
-		void WaitWithoutYield();
-		bool WaitWithoutYield(const wxTimeSpan& timeout);
-		void WaitNoCancel();
-		void WaitNoCancel(const wxTimeSpan& timeout);
-		void WaitWithoutYieldWithSpin()
-		{
-			u32 waited = 0;
-			while (true)
-			{
-				if (TryWait())
-					return;
-				if (waited >= SPIN_TIME_NS)
-					break;
-				waited += ShortSpin();
-			}
-			WaitWithoutYield();
-		}
-		bool TryWait();
-		int Count();
-
-		void Wait();
-		bool Wait(const wxTimeSpan& timeout);
-	};
-
 	class Mutex
 	{
 	protected:
@@ -502,79 +348,5 @@ namespace Threading
 	protected:
 		// Special constructor used by ScopedTryLock
 		ScopedLock(const Mutex& locker, bool isTryLock);
-	};
-
-	class ScopedTryLock : public ScopedLock
-	{
-	public:
-		ScopedTryLock(const Mutex& locker)
-			: ScopedLock(locker, true)
-		{
-		}
-		virtual ~ScopedTryLock() = default;
-		bool Failed() const { return !m_IsLocked; }
-	};
-
-	// --------------------------------------------------------------------------------------
-	//  ScopedNonblockingLock
-	// --------------------------------------------------------------------------------------
-	// A ScopedTryLock branded for use with Nonblocking mutexes.  See ScopedTryLock for details.
-	//
-	class ScopedNonblockingLock
-	{
-		DeclareNoncopyableObject(ScopedNonblockingLock);
-
-	protected:
-		NonblockingMutex& m_lock;
-		bool m_IsLocked;
-
-	public:
-		ScopedNonblockingLock(NonblockingMutex& locker)
-			: m_lock(locker)
-			, m_IsLocked(m_lock.TryAcquire())
-		{
-		}
-
-		virtual ~ScopedNonblockingLock()
-		{
-			if (m_IsLocked)
-				m_lock.Release();
-		}
-
-		bool Failed() const { return !m_IsLocked; }
-	};
-
-	// --------------------------------------------------------------------------------------
-	//  ScopedLockBool
-	// --------------------------------------------------------------------------------------
-	// A ScopedLock in which you specify an external bool to get updated on locks/unlocks.
-	// Note that the isLockedBool should only be used as an indicator for the locked status,
-	// and not actually depended on for thread synchronization...
-
-	struct ScopedLockBool
-	{
-		ScopedLock m_lock;
-		std::atomic<bool>& m_bool;
-
-		ScopedLockBool(Mutex& mutexToLock, std::atomic<bool>& isLockedBool)
-			: m_lock(mutexToLock)
-			, m_bool(isLockedBool)
-		{
-			m_bool.store(m_lock.IsLocked(), std::memory_order_relaxed);
-		}
-		virtual ~ScopedLockBool()
-		{
-			m_bool.store(false, std::memory_order_relaxed);
-		}
-		void Acquire()
-		{
-			m_lock.Acquire();
-			m_bool.store(m_lock.IsLocked(), std::memory_order_relaxed);
-		}
-		void Release()
-		{
-			m_bool.store(false, std::memory_order_relaxed);
-			m_lock.Release();
-		}
 	};
 } // namespace Threading
