+++ pcsx2/CDVD/ThreadedFileReader.cpp
@@ -16,6 +16,8 @@
 #include "PrecompiledHeader.h"
 #include "ThreadedFileReader.h"
 
+#include "common/Threading.h"
+
 // Make sure buffer size is bigger than the cutoff where PCSX2 emulates a seek
 // If buffers are smaller than that, we can't keep up with linear reads
 static constexpr u32 MINIMUM_SIZE = 128 * 1024;
