+++ pcsx2/CDVD/ThreadedFileReader.h
@@ -16,7 +16,6 @@
 #pragma once
 
 #include "AsyncFileReader.h"
-#include "common/PersistentThread.h"
 
 #include <thread>
 #include <mutex>
