+++ pcsx2/PINE.h
@@ -27,7 +27,7 @@
 #define PINE_DEFAULT_SLOT 28011
 #define PINE_EMULATOR_NAME "pcsx2"
 
-#include "common/PersistentThread.h"
+#include "gui/PersistentThread.h"
 #include "gui/SysThreads.h"
 #include <string>
 #ifdef _WIN32
