+++ pcsx2/gui/ScopedPtrMT.h
@@ -15,7 +15,7 @@
 
 #pragma once
 
-#include "Threading.h"
+#include "common/Threading.h"
 using Threading::ScopedLock;
 
 // --------------------------------------------------------------------------------------
