+++ pcsx2/gui/SysCoreThread.cpp
@@ -197,12 +197,9 @@ void SysCoreThread::ApplySettings(const Pcsx2Config& src)
 	// handle GS setting changes
 	if (GetMTGS().IsOpen() && gs_settings_changed)
 	{
-		// if by change we reopen the GS, the window handles will invalidate.
-		// so, we should block here until GS has finished reinitializing, if needed.
 		Console.WriteLn("Applying GS settings...");
 		GetMTGS().ApplySettings();
 		GetMTGS().SetVSync(EmuConfig.GetEffectiveVsyncMode());
-		GetMTGS().WaitGS();
 	}
 }
 
@@ -340,6 +337,9 @@ void SysCoreThread::TearDownSystems(SystemsMask systemsToTearDown)
 	if (systemsToTearDown & System_SPU2) SPU2close();
 	if (systemsToTearDown & System_MCD) FileMcd_EmuClose();
 
+	if (GetMTGS().IsOpen())
+		GetMTGS().WaitGS();
+
 	PerformanceMetrics::SetCPUThread(Threading::ThreadHandle());
 }
 
@@ -348,7 +348,12 @@ void SysCoreThread::OnResumeInThread(SystemsMask systemsToReinstate)
 	PerformanceMetrics::SetCPUThread(Threading::ThreadHandle::GetForCallingThread());
 	PerformanceMetrics::Reset();
 
-	GetMTGS().WaitForOpen();
+	// if GS is open, wait for it to finish (could be applying settings)
+	if (GetMTGS().IsOpen())
+		GetMTGS().WaitGS();
+	else
+		GetMTGS().WaitForOpen();
+
 	if (systemsToReinstate & System_DEV9) DEV9open();
 	if (systemsToReinstate & System_USB) USBopen(g_gs_window_info);
 	if (systemsToReinstate & System_FW) FWopen();
