+++ pcsx2/gui/pxEvents.h
@@ -19,6 +19,7 @@
 #include "common/General.h"
 #include "common/Dependencies.h"
 #include "common/Threading.h"
+#include "gui/PersistentThread.h"
 
 wxDECLARE_EVENT(pxEvt_StartIdleEventTimer, wxCommandEvent);
 wxDECLARE_EVENT(pxEvt_DeleteObject, wxCommandEvent);
